Neighbor search

Meshes.search — Function

search(pₒ, method, mask=nothing)

Return neighbors of point pₒ using method. Optionally, specify a mask for all indices of the domain.

Meshes.NeighborSearchMethod — Type

NeighborSearchMethod

A method for searching neighbors given a reference point.

Meshes.BallSearch — Type

BallSearch(domain, ball)

A method for searching neighbors in domain inside ball.

Meshes.KNearestSearch — Type

KNearestSearch(domain, k; metric=Euclidean())

A method for searching k nearest neighbors in domain according to metric.

Meshes.KBallSearch — Type

KBallSearch(domain, k, ball)

A method that searches k nearest neighbors and then filters these neighbors using a norm ball.

Meshes.BoundedSearch — Type

BoundedSearch(method, nmax)

A method for searching at most nmax neighbors using method.