Neighborhoods

Meshes.Neighborhood — Type

Neighborhood

A neighborhood is a geometry that is not attached to any specific point in the space, and is free to slide over a domain of interest.

Meshes.NormBall — Type

NormBall(radius, [metric])

A norm ball with radius and metric. Default metric is Euclidean().

Meshes.Ellipsoid — Type

Ellipsoid(semiaxes, angles; convention=TaitBryanExtr)

An ellipsoid with semiaxes and angles according to the rotation convention.

For 2D ellipses, there are two semiaxes and one rotation angle.

For 3D ellipsoids, there are three semiaxes and three rotation angles.

The list of available conventions can be found with subtypes(RotationConvention).