Points

Meshes.Point — Type

Point{Dim,T}

A point in Dim-dimensional space with coordinates of type T. The coordinates of the point provided upon construction are with respect to the canonical Euclidean basis.

Example

O = Point(0.0, 0.0) # origin of 2D Euclidean space

Notes

Type aliases are Point1, Point2, Point3, Point1f, Point2f, Point3f

Meshes.embeddim — Method

embeddim(point)

Return the number of dimensions of the space where the point is embedded.

embeddim(domain)

Return the number of dimensions of the space where the domain is embedded.

embeddim(geometry)

Return the number of dimensions of the space where the geometry is embedded.

Meshes.coordtype — Method

coordtype(point)

Return the machine type of each coordinate used to describe the point.

coordtype(domain)

Return the machine type of each coordinate used to describe the domain.

coordtype(geometry)

Return the machine type of each coordinate used to describe the geometry.

Meshes.coordinates — Method

coordinates(A::Point)

Return the coordinates of the point with respect to the canonical Euclidean basis.

Base.:- — Method

-(A::Point, B::Point)

Return the Vec associated with the direction from point A to point B.

Base.:+ — Method

+(A::Point, v::Vec)
+(v::Vec, A::Point)

Return the point at the end of the vector v placed at a reference (or start) point A.

Base.:- — Method

-(A::Point, v::Vec)
-(v::Vec, A::Point)

Return the point at the end of the vector -v placed at a reference (or start) point A.

Base.isapprox — Method

isapprox(A::Point, B::Point)

Tells whether or not the coordinates of points A and B are approximately equal.