Traits

This package defines two main traits:

Meshes.Domain — Type

Domain{Dim,T}

A domain is an indexable collection of geometries (e.g. mesh) or points For example, a collection of polygonal areas representing the states of a country can be seen as a domain.

Meshes.Data — Type

Data

A domain implementing the Domain trait together with tables of values for geometries of the domain.

See also meshdata.