Vectors

Meshes.Vec — Type

Vec{Dim,T}

A vector in Dim-dimensional space with coordinates of type T representing a direction with magnitude. A vector can be obtained by subtracting two Point objects:

Example

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

Notes

A Vec is a SVector from StaticArrays.jl

Type aliases are Vec2, Vec3, Vec2f, Vec3f