Discretization

Meshes.discretize — Function

discretize(geometry, method)

Discretize geometry with discretization method.

Meshes.DiscretizationMethod — Type

DiscretizationMethod

A method for discretizing geometries into meshes.

Meshes.FIST — Type

FIST

Fast Industrial-Strength Triangulation (FIST) of polygons.

This triangulation method is the method behind the famous Mapbox's Earcut library. It is based on a ear clipping algorithm adapted for complex n-gons with holes. It has O(n²) time complexity where n is the number of vertices. In practice it is very efficient due to heuristics implemented in the algorithm.

References

Held, M. 1998. FIST: Fast Industrial-Strength Triangulation of Polygons

Eder et al. 2018. Parallelized ear clipping for the triangulation and constrained Delaunay triangulation of polygons