Polytopes

Meshes.Polytope — Type

Polytope{N,Dim,T}

We say that a geometry is a N-polytope when it is a collection of "flat" sides that constitue a N-dimensional subspace. They are called polygon and polyhedron respectively for 2D (N=2) and 3D (N=3) subspaces, embedded in a Dim-dimensional space. The parameter N is also known as the rank or parametric dimension of the polytope: https://en.wikipedia.org/wiki/Abstract_polytope.

The term polytope expresses a particular combinatorial structure. A polyhedron, for example, can be decomposed into faces. Each face can then be decomposed into edges, and edges into vertices. Some conventions act as a mapping between vertices and higher dimensional features (edges, faces, cells...), removing the need to store all features. We follow the ordering conventions of the GMSH project: https://gmsh.info/doc/texinfo/gmsh.html#Node-ordering

Additionally, the following property must hold in order for a geometry to be considered a polytope: the boundary of a (N+1)-polytope is a collection of N-polytopes, which may have (N-1)-polytopes in common. See https://en.wikipedia.org/wiki/Polytope.

Meshes.Segment — Type

Segment(p1, p2)

An oriented line segment with end points p1, p2.

See also Line.

Meshes.Triangle — Type

Triangle(p1, p2, p3)

A triangle with points p1, p2, p3.

Meshes.Quadrangle — Type

Quadrangle(p1, p2, p3, p4)

A quadrangle with points p1, p2, p3, p4.

Meshes.Chain — Type

Chain(p1, p2, ..., pn)

A polygonal chain from a sequence of points p1, p2, ..., pn. See https://en.wikipedia.org/wiki/Polygonal_chain.

Meshes.PolyArea — Type

PolyArea(outer, [inner1, inner2, ..., innerk])

A polygonal area with outer chain, and optional inner chains inner1, inner2, ..., innerk.

Chains can be a vector of Point or a vector of tuples with coordinates for convenience.

Most algorithms assume that the outer chain is oriented counter-clockwise (CCW) and that all inner chains are oriented clockwise (CW).

Meshes.Pyramid — Type

Pyramid(p1, p2, p3, p4, p5)

A pyramid with points p1, p2, p3, p4, p5.

Meshes.Tetrahedron — Type

Tetrahedron(p1, p2, p3, p4)

A tetrahedron with points p1, p2, p3, p4.

Meshes.Hexahedron — Type

Hexahedron(p1, p2, ..., p8)

A hexahedron with points p1, p2, ..., p8.