Meshes

Types

Meshes.CartesianGrid — Type

CartesianGrid(dims, origin, spacing)

A Cartesian grid with dimensions dims, lower left corner at origin and cell spacing spacing. The three arguments must have the same length.

CartesianGrid(start, finish, dims=dims)

Alternatively, construct a Cartesian grid from a start point (lower left) to a finish point (upper right).

CartesianGrid{T}(dims)
CartesianGrid{T}(dim1, dim2, ...)

Finally, a Cartesian grid can be constructed by only passing the dimensions dims as a tuple, or by passing each dimension dim1, dim2, ... separately. In this case, the origin and spacing default to (0,0,...) and (1,1,...).

Examples

Create a 3D grid with 100x100x50 locations:

julia> CartesianGrid(100,100,50)

Create a 2D grid with 100x100 locations and origin at (10.,20.) units:

julia> CartesianGrid((100,100),(10.,20.),(1.,1.))

Create a 1D grid from -1 to 1 with 100 locations:

julia> CartesianGrid((-1.,),(1.,), dims=(100,))

Meshes.UnstructuredMesh — Type

UnstructuredMesh(points, connec)

An unstructured mesh with points and connectivities connec. The i-th element of the mesh is lazily built based on the connectivity list connec[i].

Connectivities

Meshes.Connectivity — Type

Connectivity{PL,N}

A connectivity list of N indices representing a Polytope of type PL. Indices are taken from a global vector of Point.

Connectivity objects are constructed with the connect function.

Meshes.polytopetype — Function

polytopetype(connectivity)

Return the face type (e.g. Triangle) of the connectivity.

Meshes.connect — Function

connect(list, PL)

Connect a list of indices from a global vector of Point into a Polytope of type PL.

Example

Δ = connect((1,2,3), Triangle)

Meshes.materialize — Function

materialize(connec, points)

Materialize a face using the connec list and a global vector of points.

Conventions

Meshes.GMSH — Type

Ordering convention of the GMSH project: https://gmsh.info/doc/texinfo/gmsh.html#Node-ordering